Delete questions when user initiates delete

+++ client/src/views/CompareView.vue
@@ -79,6 +79,18 @@ const deleteSubmissions = async () => {
     return;
   }
 
+  const questionsDeleteResponse = await fetch(questionsEndpoint(sessionInfo.value?.questions), {
+    method: "DELETE",
+  });
+
+  if (questionsDeleteResponse.status !== 204) {
+    const { error } = await questionsDeleteResponse.json();
+
+    toast.add({ severity: "error", summary: "Error", detail: error, life: ERROR_TOAST_TTL });
+
+    return;
+  }
+
   const sessionDeleteResponse = await fetch(sessionsEndpoint(sharingCode.value), {
     method: "DELETE",
   });
